backend-python/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
pdfplumber==0.10.3
python-docx==0.8.11
docx2txt==0.8
mammoth==1.6.0
pandas==2.1.0
openai==1.3.0
xlsxwriter==3.1.9
sentence-transformers==2.2.2
chromadb==0.4.18
numpy==1.24.3
python-dotenv==1.0.0
requests==2.31.0
